+++ apps/commerce-api/src/main/kotlin/com/loopers/domain/product/Product.kt
@@ -65,7 +65,7 @@ class Product internal constructor(
 
     fun isDeleted(): Boolean = deletedAt != null
 
-    fun update(name: String, price: Int, salesStatus: SalesStatus) {
+    fun update(name: String, price: Long, salesStatus: SalesStatus) {
         this.name = ProductName.of(name)
         this.price = ProductPrice.of(price)
         this.salesStatus = salesStatus
@@ -74,7 +74,7 @@ class Product internal constructor(
     companion object {
         fun create(
             name: String,
-            price: Int,
+            price: Long,
             stock: Int,
             likeCount: Long,
             brandId: Long,
